RavenDB-9487 - better calculation of low memory
